Dumped the pug view system.
Swapped the AppFormPopup while also removing the progress bar.
While fixing the redirect feature and also message not showing up.

Made some changes on url in '/sign_in' page to help the issue in AppFormPopup.

Tried to mitigate the error in '/author' route but havent got achieved so.

renamed the book creation/modification handler from 'createBook' to 'bookHandler'

Added new auth method in authHandler for checking the changes done on book are done by author or not.